requirements.txt
# Python 依赖列表
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
requests>=2.28.0